+++ 0x04-more_functions_nested_loops/main.h
@@ -23,6 +23,6 @@ void print_square(int);
 
 void print_triangle(int);
 
-void print_number(long int);
+void print_number(int);
 
 #endif
